requirements.txt
django-restql
django
djangorestframework
djangorestframework-simplejwt
django-filter
requests
bs4
